+++ backend/src/main/java/com/bada/cali/controller/BasicController.java
@@ -56,7 +56,7 @@ public class BasicController {
 	
 	// 첨부파일 리스트 [모달]
 	@PostMapping(value = "/fileList")
-	public String fileList(Model model, @RequestParam String refTableName, @RequestParam int refTableId) {
+	public String fileList(Model model, @RequestParam String refTableName, @RequestParam Long refTableId) {
 		
 		// 작업자까지 join을 통해서 가져온다.
 		List<FileInfoDTO.FileListRes> fileList = fileService.getFileInfosWithJoin(refTableName, refTableId);
