+++ backend/src/main/java/com/bada/cali/repository/AgentRepository.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 
-public interface AgentRepository extends JpaRepository<Agent, Integer> {
+public interface AgentRepository extends JpaRepository<Agent, Long> {
 	
 	@Query("""
 			SELECT a
@@ -43,10 +43,10 @@ public interface AgentRepository extends JpaRepository<Agent, Integer> {
 	);
 	
 	// 단건 조회
-	Agent findByIsVisibleAndId(YnType isVisible, Integer id);
+	Agent findByIsVisibleAndId(YnType isVisible, Long id);
 	
 	// 삭제대상 업체정보
-	List<Agent> findAllByIdInAndIsVisible(Collection<Integer> ids, YnType isVisible);
+	List<Agent> findAllByIdInAndIsVisible(Collection<Long> ids, YnType isVisible);
 	
 	// 기본: 삭제된 것을 제외하고 전체 가져오기
 	Page<Agent> findByIsVisible(YnType isVisible, Pageable pageable);
@@ -62,10 +62,10 @@ public interface AgentRepository extends JpaRepository<Agent, Integer> {
 			              a.deleteMemberId = :deleteMemberId
 			        where a.id in :agentIds
 			""")
-	int delAgentByIds(@Param("agentIds") Collection<Integer> agentIds,
+	int delAgentByIds(@Param("agentIds") Collection<Long> agentIds,
 					  @Param("isVisible") YnType isVisible,
 					  @Param("deleteDatetime") LocalDateTime deleteDatetime,
-					  @Param("deleteMemberId") Integer deleteMemberId
+					  @Param("deleteMemberId") Long deleteMemberId
 	);
 	
 	// 존재하는 업체 그룹명들 반환 (null과 빈문자열 모두 배제하기 위해 JPQL 작성 && 중복그룹명 제외 by distinct)
@@ -87,7 +87,7 @@ public interface AgentRepository extends JpaRepository<Agent, Integer> {
 						WHERE a.id in :ids
 			""")
 	int updateAgentGroupName(
-			@Param("ids") List<Integer> ids,
+			@Param("ids") List<Long> ids,
 			@Param("groupName") String groupName
 	);
